% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rePath.R
\name{rePath}
\alias{rePath}
\title{Convert a Windows path on the clipboard to be R-friendly.}
\usage{
rePath()
}
\value{
An R-friendly Windows path
}
\description{
Convert a Windows path on the clipboard to be R-friendly.
}
